my-chat-ui/sovits-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pyyaml==6.0.1
numpy==1.24.3
torch==2.1.0
torchaudio==2.1.0
librosa==0.10.1
soundfile==0.12.1
requests==2.31.0
python-multipart==0.0.6
pydub==0.25.1
websockets==12.0
aiohttp==3.9.1
matplotlib==3.8.0
pandas==2.0.3
pillow==10.0.0
